# ✅ Implement TDD GREEN Phase: Core Models & Safety Validation (100% Complete)