--- a/Cargo.toml
+++ b/Cargo.toml
@@ -55,6 +55,7 @@ chrono = { version = "0.4", features = ["serde"] }
 
 # Regex for safety validation
 regex = "1"
+once_cell = "1.19"
 
 # Async trait support
 async-trait = "0.1"