… .claude/commands/analyze.md
… .claude/commands/clarify.md
… .claude/commands/constitution.md
… .claude/commands/implement.md
… .config/nextest.toml
… .cursor/commands/analyze.md
… .cursor/commands/clarify.md
… .cursor/commands/constitution.md
… .cursor/commands/implement.md
… .cursor/commands/plan.md
… .cursor/commands/specify.md
… .cursor/commands/tasks.md
… .specify/memory/constitution.md
… .specify/scripts/bash/update-agent-context.sh
… 2025-10-01-Phase2.txt
… 2025-10-01-TDD-Green-moduleB.txt
… 2025-10-01-check-the-pr.txt
… 2025-10-01-improve-test-runner.txt
… 2025-10-01-vscode-run-tests.txt
… Cargo.lock
… exports/README.md
… exports/session-logs/2025-09-12-caveat-the-messages-below-were-generated-by-the-u.txt
… exports/session-logs/2025-09-13-agents.txt
… exports/session-logs/2025-09-13-caveat-the-messages-below-were-generated-by-the-u.txt
… exports/session-logs/2025-09-13-gitignore.txt
… exports/session-logs/2025-09-13-readme.txt
… exports/session-logs/2025-10-01-Phase2.txt
… exports/session-logs/2025-10-01-TDD-Green-moduleB.txt
… exports/session-logs/2025-10-01-check-the-pr.txt
… exports/session-logs/2025-10-01-green-moduleB-review.txt